agent/requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
faiss-cpu>=1.9.0
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
